requirements/dev.txt
# Additional requirements useful for development.

-r base.txt

mock==0.7.0rc1
coverage==3.4

Fabric==1.0.1
pycrypto==2.3
paramiko==1.7.7.1

epio==0.2.7
httplib2==0.6.0

ipython==0.10.2

django-webtest==1.3
WebTest==1.2.3
WebOb==1.0.7
